+++ src/components/POLine/RoutingList/RoutingListUsers/RoutingListUsers.js
@@ -2,31 +2,35 @@ import {
   useCallback,
   useMemo,
 } from 'react';
-import {
-  keyBy,
-  map,
-} from 'lodash';
+import keyBy from 'lodash/keyBy';
+import map from 'lodash/map';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 
 import {
+  DragDropMCL,
   useToggle,
   useUsersBatch,
 } from '@folio/stripes-acq-components';
 import {
-  Pluggable,
-  useStripes,
-} from '@folio/stripes/core';
-import {
   Button,
   Col,
   ConfirmationModal,
-  List,
+  Layout,
   Loading,
   Row,
 } from '@folio/stripes/components';
+import {
+  Pluggable,
+  useStripes,
+} from '@folio/stripes/core';
 
-import { RoutingListUserItem } from './RoutingListUserItem';
+import {
+  ROUTING_LIST_USERS_COLUMN_MAPPING,
+  VISIBLE_COLUMNS,
+  columnWidths,
+} from './constants';
+import { getRoutingListUsersFormatter } from './utils';
 
 import css from './RoutingListUsers.css';
 
@@ -59,14 +63,17 @@ export const RoutingListUsers = ({
   }, [onAddUsers, toggleUnassignUsersModal]);
 
   const selectedUsersMap = useMemo(() => (ids?.length ? keyBy(users, 'id') : {}), [users, ids]);
+  const orderedUsersList = useMemo(() => {
+    return ids?.map((id) => selectedUsersMap[id]);
+  }, [ids, selectedUsersMap]);
 
-  const renderItem = useCallback((user) => (
-    <RoutingListUserItem
-      canRemove={editable}
-      onRemove={onRemoveUser}
-      user={user}
-    />
-  ), [editable, onRemoveUser]);
+  const onUpdate = (newLinesList) => {
+    const updatedLinesList = newLinesList.map(({ id }) => id);
+
+    onAddUsers(updatedLinesList);
+  };
+
+  const formatter = getRoutingListUsersFormatter({ onRemoveUser, editable });
 
   if (isLoading) return <Loading />;
 
@@ -74,18 +81,23 @@ export const RoutingListUsers = ({
     <>
       <Row>
         <Col xs={12}>
-          <List
-            items={ids?.length ? users : []}
-            listStyle="default"
-            marginBottom0
-            itemFormatter={renderItem}
+          <DragDropMCL
+            contentData={orderedUsersList}
+            columnMapping={ROUTING_LIST_USERS_COLUMN_MAPPING}
+            formatter={formatter}
+            id="users-draggable-list"
+            loading={isLoading}
+            onUpdate={onUpdate}
             isEmptyMessage={<FormattedMessage id="ui-orders.routing.list.users.empty" />}
+            visibleColumns={VISIBLE_COLUMNS}
+            isRowDraggable={() => editable}
+            columnWidths={columnWidths}
           />
         </Col>
       </Row>
       {
         editable && (
-          <div className={css.actions}>
+          <Layout className={`flex full ${css.actions}`}>
             <Pluggable
               aria-haspopup="true"
               dataKey="users"
@@ -108,7 +120,7 @@ export const RoutingListUsers = ({
             >
               <FormattedMessage id="ui-orders.routing.list.removeUsers" />
             </Button>
-          </div>
+          </Layout>
         )
       }